pyproject.toml
[project]
name = "stewart-filmscreen-homeassistant"
version = "0.1.0"
description = "Home Assistant integration for Stewart Filmscreen CVM"
requires-python = ">=3.11"
dependencies = [
  "stewart-filmscreen>=0.1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "pytest-homeassistant-custom-component>=0.13.0",
  "ruff>=0.8.0",
  "ty",
]

[tool.uv]
package = false
sources = { stewart-filmscreen = { path = "../py-stewart-filmscreen", editable = true } }

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
  "--cov=custom_components/stewart_filmscreen",
  "--cov-report=term-missing",
  "--cov-report=html",
  "-v",
]
filterwarnings = [
  "ignore::DeprecationWarning",
]
markers = [
  "integration_real: read-only tests against a real Stewart CVM (manual opt-in via env vars)",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]
